refactor: separate into hooks + styling updates
