+++ apps/mobile/src/app/(tabs)/create/index.tsx
@@ -4,202 +4,48 @@ import { CreateCameraDrawer } from '@/features/create/components/camera/drawer';
 import { CreateModePanel } from '@/features/create/components/mode-panel';
 import { CreateModeToggle } from '@/features/create/components/mode-toggle';
 import { CreateSubmitButton } from '@/features/create/components/submit-button';
-import {
-  CREATE_CAMERA_REVEAL_RESET_THRESHOLD,
-  CREATE_CAMERA_REVEAL_THRESHOLD,
-} from '@/features/create/constants';
-import type { CreateFormValues, CreateMode, CreateParams } from '@/features/create/types';
-import { getModeParam, getStringParam, toFileUri } from '@/features/create/utils';
+import { CREATE_DRAWER_SNAP_POINTS } from '@/features/create/constants';
+import { useCreateDrawer } from '@/features/create/hooks/use-create-drawer';
+import { useCreateForm } from '@/features/create/hooks/use-create-form';
+import { useCreateMode } from '@/features/create/hooks/use-create-mode';
 import { GuestMode } from '@/features/profile/components/guest-mode';
-import { setTabBarHidden } from '@/lib/tab-bar-visibility';
-import { api } from '@fomo/backend/convex/_generated/api';
 import { useBottomTabBarHeight } from '@react-navigation/bottom-tabs';
-import { useIsFocused } from '@react-navigation/native';
-import { useQuery } from 'convex/react';
-import { useLocalSearchParams } from 'expo-router';
-import { useCallback, useEffect, useMemo, useState } from 'react';
-import { useForm } from 'react-hook-form';
-import { Alert, ScrollView, useWindowDimensions, View } from 'react-native';
-import { Gesture, GestureDetector } from 'react-native-gesture-handler';
-import Animated, {
-  Easing,
-  useAnimatedReaction,
-  useAnimatedStyle,
-  useSharedValue,
-  withTiming,
-} from 'react-native-reanimated';
-
+import { ScrollView, View } from 'react-native';
+import { GestureDetector } from 'react-native-gesture-handler';
+import Animated from 'react-native-reanimated';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
-import { scheduleOnRN } from 'react-native-worklets';
 
 export default function CreateScreen() {
-  const params = useLocalSearchParams<CreateParams>();
   const tabBarHeight = useBottomTabBarHeight();
-  const isFocused = useIsFocused();
-  const { height, width } = useWindowDimensions();
   const insets = useSafeAreaInsets();
-  const allTags = useQuery(api.tags.getAllTags) ?? [];
-  const drawerAnimatedIndex = useSharedValue(0);
-  const drawerAnimatedPosition = useSharedValue(0);
-  const isCameraRevealed = useSharedValue(false);
-  const drawerSnapPoints = useMemo(() => ['20.5%', '100%'], []);
-  const { control, handleSubmit, setValue } = useForm<CreateFormValues>({
-    defaultValues: {
-      post: {
-        description: '',
-        tags: [],
-        media: {
-          uri: '',
-          type: undefined,
-        },
-      },
-      event: {
-        description: '',
-        tags: [],
-        media: {
-          uri: '',
-          type: undefined,
-        },
-      },
-    },
-  });
 
-  const [selectedMode, setSelectedMode] = useState<CreateMode>(getModeParam(params.mode));
-  const [drawerIndex, setDrawerIndex] = useState(0);
-  const [isTagMenuOpen, setIsTagMenuOpen] = useState(false);
-  const modeProgress = useSharedValue(getModeParam(params.mode) === 'event' ? 1 : 0);
+  const {
+    selectedMode,
+    isEventMode,
+    modeProgress,
+    mediaHeight,
+    contentWidth,
+    setMode,
+    modeSwipeGesture,
+    contentTrackStyle,
+    postPanelStyle,
+    eventPanelStyle,
+  } = useCreateMode();
 
-  const mediaUriParam = getStringParam(params.mediaUri);
-  const mediaTypeParam = getStringParam(params.mediaType);
-  const incomingMode = getModeParam(params.mode);
+  const {
+    isFocused,
+    drawerIndex,
+    drawerAnimatedIndex,
+    drawerAnimatedPosition,
+    openCamera,
+    handleDrawerChange,
+    handleCloseDrawer,
+  } = useCreateDrawer();
 
-  useEffect(() => {
-    setSelectedMode(getModeParam(params.mode));
-  }, [params.mode]);
-
-  useEffect(() => {
-    modeProgress.value = withTiming(selectedMode === 'event' ? 1 : 0, {
-      duration: 220,
-      easing: Easing.out(Easing.cubic),
-    });
-  }, [modeProgress, selectedMode]);
-
-  useEffect(() => {
-    if (!mediaUriParam) {
-      return;
-    }
-
-    const nextMedia = {
-      uri: toFileUri(mediaUriParam),
-      type: mediaTypeParam,
-    };
-
-    setValue(incomingMode === 'event' ? 'event.media' : 'post.media', nextMedia, {
-      shouldDirty: true,
-    });
-
-    setDrawerIndex(0);
-  }, [incomingMode, mediaTypeParam, mediaUriParam, setValue]);
-
-  useEffect(() => {
-    setTabBarHidden(drawerIndex > 0);
-  }, [drawerIndex]);
-
-  useEffect(() => {
-    if (!isFocused && drawerIndex === 0) {
-      setTabBarHidden(false);
-      return;
-    }
-
-    if (!isFocused && drawerIndex > 0) {
-      setTabBarHidden(true);
-    }
-  }, [drawerIndex, isFocused]);
-
-  const isEventMode = selectedMode === 'event';
-  const mediaHeight = Math.max(180, Math.min(250, height * 0.28));
-  const contentWidth = Math.max(width - 32, 0);
-
-  const setMode = (mode: CreateMode) => {
-    if (mode === selectedMode) {
-      return;
-    }
-
-    setSelectedMode(mode);
-  };
-
-  const contentTrackStyle = useAnimatedStyle(
-    () => ({
-      transform: [{ translateX: -contentWidth * modeProgress.value }],
-    }),
-    [contentWidth]
+  const { control, setValue, onSubmit, allTags, isTagMenuOpen, setIsTagMenuOpen } = useCreateForm(
+    selectedMode,
+    handleCloseDrawer
   );
-
-  const switchModeFromSwipe = useCallback(
-    (translationX: number, velocityX: number) => {
-      const hasEnoughDistance = Math.abs(translationX) >= 48;
-      const hasEnoughVelocity = Math.abs(velocityX) >= 700;
-
-      if (!hasEnoughDistance && !hasEnoughVelocity) {
-        return;
-      }
-
-      if (translationX < 0 && selectedMode !== 'event') {
-        setMode('event');
-      }
-
-      if (translationX > 0 && selectedMode !== 'post') {
-        setMode('post');
-      }
-    },
-    [selectedMode]
-  );
-
-  const openCamera = useCallback(() => {
-    setDrawerIndex(1);
-  }, []);
-
-  const handleDrawerChange = useCallback((nextIndex: number) => {
-    setDrawerIndex(Math.max(nextIndex, 0));
-  }, []);
-
-  const handleCloseDrawer = useCallback(() => {
-    setDrawerIndex(0);
-  }, []);
-
-  useAnimatedReaction(
-    () => drawerAnimatedIndex.value,
-    (currentIndex) => {
-      if (!isCameraRevealed.value && currentIndex >= CREATE_CAMERA_REVEAL_THRESHOLD) {
-        isCameraRevealed.value = true;
-        scheduleOnRN(setTabBarHidden, true);
-      } else if (isCameraRevealed.value && currentIndex <= CREATE_CAMERA_REVEAL_RESET_THRESHOLD) {
-        isCameraRevealed.value = false;
-        scheduleOnRN(setTabBarHidden, false);
-      }
-    },
-    []
-  );
-
-  const onSubmit = handleSubmit((values) => {
-    const activeValues = values[selectedMode];
-    const normalizedDescription = activeValues.description.trim();
-    const normalizedTags = activeValues.tags.join(', ');
-    const noun = isEventMode ? 'Event' : 'Post';
-
-    Alert.alert(
-      `${noun} Data`,
-      `Description: ${normalizedDescription || '(none)'}\nTags: ${normalizedTags || '(none)'}`,
-      [{ text: 'OK' }]
-    );
-  });
-
-  const modeSwipeGesture = Gesture.Pan()
-    .activeOffsetX([-30, 30])
-    .failOffsetY([-20, 20])
-    .onEnd((event) => {
-      scheduleOnRN(switchModeFromSwipe, event.translationX, event.velocityX);
-    });
 
   return (
     <Screen>
@@ -209,7 +55,7 @@ export default function CreateScreen() {
 
       <Authenticated>
         <GestureDetector gesture={modeSwipeGesture}>
-          <View className="flex-1 bg-background">
+          <View className="flex-1 bg-surface-muted">
             <CreateModeToggle
               selectedMode={selectedMode}
               modeProgress={modeProgress}
@@ -219,57 +65,50 @@ export default function CreateScreen() {
             <ScrollView
               style={{ flex: 1 }}
               contentContainerStyle={{
-                paddingBottom: tabBarHeight + (isTagMenuOpen ? 320 : 176),
+                paddingBottom: tabBarHeight + (isTagMenuOpen ? 220 : 176),
                 paddingHorizontal: 16,
                 rowGap: 18,
               }}
               showsVerticalScrollIndicator={false}
               keyboardShouldPersistTaps="handled"
             >
-              <View style={{ overflow: 'hidden' }}>
-                <Animated.View
-                  className="flex-row items-start"
-                  style={[
-                    contentTrackStyle,
-                    {
-                      width: contentWidth * 2,
-                    },
-                  ]}
-                >
-                  <View style={{ width: contentWidth }}>
-                    <CreateModePanel
-                      control={control}
-                      setValue={setValue}
-                      mode="post"
-                      mediaHeight={mediaHeight}
-                      isTagMenuOpen={!isEventMode && isTagMenuOpen}
-                      setIsTagMenuOpen={setIsTagMenuOpen}
-                      allTags={allTags}
-                      openCamera={openCamera}
-                    />
-                  </View>
-
-                  <View style={{ width: contentWidth }}>
-                    <CreateModePanel
-                      control={control}
-                      setValue={setValue}
-                      mode="event"
-                      mediaHeight={mediaHeight}
-                      isTagMenuOpen={isEventMode && isTagMenuOpen}
-                      setIsTagMenuOpen={setIsTagMenuOpen}
-                      allTags={allTags}
-                      openCamera={openCamera}
-                    />
-                  </View>
+              <Animated.View
+                className="flex-row items-start"
+                style={[contentTrackStyle, { width: contentWidth * 2 }]}
+              >
+                <Animated.View style={[{ width: contentWidth }, postPanelStyle]}>
+                  <CreateModePanel
+                    control={control}
+                    setValue={setValue}
+                    mode="post"
+                    mediaHeight={mediaHeight}
+                    isTagMenuOpen={!isEventMode && isTagMenuOpen}
+                    setIsTagMenuOpen={setIsTagMenuOpen}
+                    allTags={allTags}
+                    openCamera={openCamera}
+                  />
                 </Animated.View>
-              </View>
+
+                <Animated.View style={[{ width: contentWidth }, eventPanelStyle]}>
+                  <CreateModePanel
+                    control={control}
+                    setValue={setValue}
+                    mode="event"
+                    mediaHeight={mediaHeight}
+                    isTagMenuOpen={isEventMode && isTagMenuOpen}
+                    setIsTagMenuOpen={setIsTagMenuOpen}
+                    allTags={allTags}
+                    openCamera={openCamera}
+                  />
+                </Animated.View>
+              </Animated.View>
             </ScrollView>
           </View>
         </GestureDetector>
 
         <CreateCameraDrawer
           drawerIndex={drawerIndex}
-          drawerSnapPoints={drawerSnapPoints}
+          drawerSnapPoints={CREATE_DRAWER_SNAP_POINTS}
           mode={selectedMode}
           isParentFocused={isFocused}
           animatedIndex={drawerAnimatedIndex}
